Idea
Made a possible solution/base idea for a forum in java. Created a Profile object, Credentials object, Users object to hold all created profiles, a Post object (unimplemented), and a Main driver method that acts as the "forum". We could move all methods into a seperate package for the forum as all the code is under the run() method.